requirements.txt
gradio
python-dotenv
langchain
langchain-pinecone
langchain-huggingface
langchain-google-genai
langchain-community
langchain_core